Feature/eth vrf integration tests (#9708)
* integration tests for VRF V2 Plus. WIP

* support vrfv2 plus in bhs and bhf

* small fixes

* fix failing force fulfill test cases

* make changes to vrfv2 consumer contract to make batching tests pass

* fix import

* add ci profile

* fix v1 test

* prettier

* Update contracts/src/v0.8/vrf/testhelpers/VRFV2PlusExternalSubOwnerExample.sol


* Update contracts/src/v0.8/vrf/testhelpers/VRFV2PlusRevertingExample.sol


* Update contracts/src/v0.8/vrf/testhelpers/VRFV2PlusSingleConsumerExample.sol


* address comments

---------

+++ contracts/src/v0.8/vrf/VRFV2PlusSubscriptionManager.sol
@@ -114,7 +114,7 @@ contract VRFV2PlusSubscriptionManager is ConfirmedOwner {
     // add the consumers to the subscription on the new coordinator
     // and set the new coordinator on the consumers
     // note that this is bounded by MAX_CONSUMERS in the coordinator
-    for (uint i = 0; i < consumers.length; i++) {
+    for (uint256 i = 0; i < consumers.length; i++) {
       newCoord.addConsumer(newSubId, consumers[i]);
       IVRFMigratableConsumerV2Plus(consumers[i]).setVRFCoordinator(newCoordinator);
     }
